+++ core-ng/src/main/java/core/framework/api/module/LogConfig.java
@@ -2,8 +2,7 @@ package core.framework.api.module;
 
 import core.framework.api.log.MessageFilter;
 import core.framework.impl.log.ActionLogger;
-import core.framework.impl.log.KafkaLogForwarder;
-import core.framework.impl.log.RabbitMQLogForwarder;
+import core.framework.impl.log.LogForwarder;
 import core.framework.impl.log.TraceLogger;
 import core.framework.impl.log.stat.CollectStatTask;
 import core.framework.impl.module.ModuleContext;
@@ -57,20 +56,11 @@ public final class LogConfig {
         }
     }
 
-    public void forwardLogToRabbitMQ(String host) {
+    public void forwardLog(String kafkaURI) {
         if (context.isTest()) {
             logger.info("disable log forwarding during test");
         } else {
-            context.logManager.logForwarder = new RabbitMQLogForwarder(host, context.logManager.appName);
-            context.backgroundTask().scheduleWithFixedDelay(new CollectStatTask(context.logManager.logForwarder), Duration.ofSeconds(10));
-        }
-    }
-
-    public void forwardLogToKafka(String uri) {
-        if (context.isTest()) {
-            logger.info("disable log forwarding during test");
-        } else {
-            context.logManager.logForwarder = new KafkaLogForwarder(uri, context.logManager.appName);
+            context.logManager.logForwarder = new LogForwarder(kafkaURI, context.logManager.appName);
             context.backgroundTask().scheduleWithFixedDelay(new CollectStatTask(context.logManager.logForwarder), Duration.ofSeconds(10));
         }
     }