+++ core-ng/src/main/java/core/framework/api/module/SystemModule.java
@@ -53,8 +53,7 @@ public final class SystemModule extends Module {
                 log().writeTraceLogToFile(Paths.get(path));
             }
         });
-        property("sys.log.remoteLogHost").ifPresent(host -> log().forwardLogToRabbitMQ(host));
-        property("sys.log.kafkaURI").ifPresent(uri -> log().forwardLogToKafka(uri));
+        property("sys.log.kafkaURI").ifPresent(uri -> log().forwardLog(uri));
 
         property("sys.rabbitMQ.host").ifPresent(hosts -> queue().hosts(Strings.split(hosts, ',')));
         property("sys.rabbitMQ.user").ifPresent(user -> queue().user(user));