+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 ## Change log
 ### 4.8.3 (10/27/2016 - )
 * es: update es to 5.0.0
+* log: removed rabbitmq log forward support
 
 ### 4.8.2 (10/26/2016)
 * log: rename kafkaHost to kafkaURI